Rename to KeywordsChipsInput
RISDEV-1068
